move file filters to files directory
